Separate the persisters again
